pyproject.toml
[project]
name = "monet"
version = "0.1.0"
description = ""
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Evan O'Keefe", email = "evan.okeefe39@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "sqlalchemy>=2.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "langgraph>=0.2",
]

[project.urls]
Homepage = "https://github.com/evanokeefe39/monet"
Repository = "https://github.com/evanokeefe39/monet"
Issues = "https://github.com/evanokeefe39/monet/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/monet"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "hypothesis>=6.0",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "langgraph>=0.2",
    "ruff>=0.9",
    "mypy>=1.14",
    "mkdocs-material>=9.7.6",
    "pre-commit>=4.5.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "UP",
    "B",
    "SIM",
    "TCH",
    "RUF",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = false
